Move template specific components to templates directory
