+++ gatsby/src/templates/ApiReference/Schema/Content/Content.tsx
