+++ gatsby/src/templates/ApiReference/Schema/Content/Properties.module.scss
