+++ gatsby/src/templates/ApiReference/Schema/Content/index.ts
