+++ gatsby/src/templates/ApiReference/Schema/Heading.module.scss
