+++ gatsby/src/templates/ApiReference/Schema/Heading.tsx
