+++ gatsby/src/templates/ApiReference/Schema/Type.module.scss
