+++ gatsby/src/templates/ApiReference/Schema/Type.tsx
