+++ gatsby/src/templates/Page/TableOfContents.module.scss
