+++ gatsby/src/templates/Page/index.ts
@@ -1,0 +1,1 @@
+export { default as default } from './Page';
